Update ehsensibleautodownloads.user.js
Added optional integration with E-H Visited script, so that downloaded galleries count as viewed. Barring any surprising changes to the storage format of that script, this should be compatible with future updates with that script. To be safe, though, you may want to export the list of visited sites from time to time.